refactor: Recruit Application 도메인 분리
