+++ src/main/java/com/ssafy/ssafsound/domain/recruitapplication/domain/MatchStatus.java
@@ -1,4 +1,4 @@
-package com.ssafy.ssafsound.domain.recruit.domain;
+package com.ssafy.ssafsound.domain.recruitapplication.domain;
 
 public enum MatchStatus {
     WAITING_REGISTER_APPROVE, WAITING_APPLICANT, DONE, REJECT
